PMIS_Assistant/requirements.txt
# GUI Framework
PyQt6>=6.6.0

# Database
psycopg[binary]>=3.1.0
SQLAlchemy==2.0.25

# AI and Machine Learning
requests==2.31.0
python-dotenv==1.0.0
google-genai==0.3.0

# Image Processing
Pillow>=10.4.0
pytesseract==0.3.10

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Windows-specific
pywin32>=306
winshell==0.6

# System Monitoring
pynput==1.7.6

# File and Path Handling
pathlib2==2.3.7

# Date and Time
python-dateutil==2.8.2

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-qt==4.2.0

# Development
black==23.11.0
flake8==6.1.0

# Optional: For better OCR performance (requires additional installation)
# opencv-python==4.8.1.78

# Optional: For enhanced text processing
# nltk==3.8.1
# spacy==3.7.2